Tidied up to suppress Eclipse nagging
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/branches/FOP_0-20-0_Alt-Design@197445 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/fonts/truetype/FontFileReader.java
@@ -161,9 +161,9 @@ public class FontFileReader {
         final byte buf = read();
 
         if (buf < 0) {
-            return (int)(256 + buf);
+            return 256 + buf;
         } else {
-            return (int)buf;
+            return buf;
         }
     }
 
@@ -187,7 +187,7 @@ public class FontFileReader {
      */
     public final int readTTFUShort() throws IOException {
         final int ret = (readTTFUByte() << 8) + readTTFUByte();
-        return (int)ret;
+        return ret;
     }
 
     /**
