+++ src/java/org/apache/fop/fonts/type1/PFMFile.java
@@ -164,10 +164,10 @@ public class PFMFile extends AbstractLogEnabled {
 
         getLogger().info(i + " kerning pairs");
         while (i > 0) {
-            int g1 = (int)inStream.readByte();
+            int g1 = inStream.readByte();
             i--;
 
-            int g2 = (int)inStream.readByte();
+            int g2 = inStream.readByte();
 
             int adj = inStream.readShort();
             if (adj > 0x8000) {
